Hoist RouteCalculation class to its own file
The original TradeCalculation.py file contained six classes - adding
to its complexity and thus confusion it caused.  As the first step
in breaking up that monstrosity, out goes RouteCalculation to its
own file.
